Add CounterComponent, HomeComponent and NavMenuComponent UIs
Created user interfaces for CounterComponent, HomeComponent, and NavMenuComponent. CounterComponent is a simple counter that increments on button click. HomeComponent displays a list of recently added boxes and some statistical data. NavMenuComponent is a navigation bar. Also, Apexcharts and ng-apexcharts are added for charting purposes in HomeComponent. UI is stylized using Tailwind CSS with DaisyUI for a simple, clean design. Changes are not merely cosmetic; they aim at user-friendliness and app's overall functionality.