Remove Counter feature, add BoxList and new interfaces
This commit removes the unused Counter feature and integrates a new component, BoxList, showing a list of boxes.

The changes include:

- Deletion of Counter related files. No usage across the app found.
- Addition of BoxList component, with corresponding html templates, spec and service to handle HTTP calls
- Creation of several new interfaces, including Box, Customer, Dimensions, and Orders to support the BoxList component
- An update to the AppModule to include HttpClientModule and the new BoxListModule.
- Update of tailwind.config.js, changing the theme from 'cupcake' to 'emerald'.
- Inclusion of a new Createbox component and service to handle the creation of new boxes.

The reason behind these changes was to clean up the previous unused code and introduce the new BoxList component to the application.

Please note that, this commit entails many changes that could have been split into smaller commits.