"Refactored order creation functionality and added address integration
The order creation functionality was refactored to incorporate customer address details in the order form. Separate components were created for order creation and box creation; earlier both were under one component. Also deleted unused `CreateboxComponent`. Additionally, ShippingStaus was introduced in the Order interface to track the shipping status of an order. Revamped the lists view of orders to display customer's email and shipping status. This was done to enhance the order creation feature, providing a more comprehensive form for users when placing orders and to improve the ease of tracking orders."