Update the Box Update functionality
Modified several files to fix and improve the Box Update feature in the application. First, the URL routing was updated to include 'boxId' as a variable, which is necessary to identify which box is being updated.

In 'box-interface.ts', the Box Interface was updated by replacing 'dimensions' with 'dimensionsDto' and adding 'stock' and 'price' properties. This change aligns it with the new model schema for boxes and supports new features of a box.

The 'box-service.ts' was changed to return a Promise using 'firstValueFrom()' in the 'getbyId()' function. This ensures that there's no unhandled Observable after the process, guarding against potential memory leaks.

In 'updatebox.component.html' and 'updatebox.component.ts', the component was changed from creating a box to updating an existing one. This was done by updating the form fields from hard coded to binding the box values retrieved via the 'boxId' from the URL.

The result is a more robust and reliable Box Update feature.