Add shipping status change functionality in orders list
A dropdown menu has been added to each order in the orders list that allows the user to update the shipping status of the order. This was done to provide a straightforward way for users to manage order statuses. The necessary methods for updating the shipping status have been included in the order service, and the ShippingStatus interface has also been adjusted to reflect the new changes.