Remove OrderRowComponent and add Simpsons image to customer
OrderRowComponent was removed from the project due to its redundancy. The functionality that was supposed to be done by OrderRowComponent is now executed directly in HomeComponent. The Simpsons image was added to the customer which is randomly assigned during order creation on the backend in OrderRepository.cs for a fun feature in the app layout. Changes were propagated across related frontend files and the necessary property simpsonImgUrl was added to customer-interface.ts and CreateCustomerDto.cs.