Reduce generated data and fix order calculation
Reduced generated data's range when initializing the database, to ease testing and improve performance. Changed orders creation to be based on the current month rather than a fixed value, and reduced orders and boxes values.

Changes in the HTML file was for displaying prices with fixed decimal places to enhance visual consistency.

Modified the customer repository: when there's a conflict on customer's email, SQL command now attempts to select and return existing customer data instead of failing. This mitigates the potential failure when trying to commit data.